return 'configured' and 'valid' state for keys
