+++ lib/ws_servers/api/validate_modes.js
@@ -1,50 +1,30 @@
 'use strict'
 
-const { RESTv2 } = require('bfx-api-node-rest')
 const send = require('../../util/ws/send')
-const decryptAPICredentials = require('../../util/decrypt_api_credentials')
+const validateCredentials = require('../../util/validate_credentials')
 
-async function validateKeys (url, apiKey, apiSecret) {
-  const rest = new RESTv2({
-    url,
-    apiKey,
-    apiSecret,
-    transform: true
-  })
-
-  try {
-    await rest.userInfo()
-    return true
-  } catch (e) {
-    const isKeyInvalid = e.error[2].match(/apikey:.*invalid/)
-
-    if (isKeyInvalid) return false
-    throw e
-  }
-}
-
-async function validateCredentials (db, authPassword, mode) {
+async function getCredentials (db, mode) {
   const { Credential } = db
   const [credentials] = await Credential.find([['mode', '=', mode]])
 
-  if (!credentials) {
-    return false
+  return credentials
+}
+
+async function getCredentialsState (db, mode, authPassword) {
+  const credentials = await getCredentials(db, mode)
+
+  return {
+    configured: Boolean(credentials),
+    valid: await validateCredentials(credentials, authPassword)
   }
-
-  const { key, secret } = await decryptAPICredentials({
-    password: authPassword,
-    credentials
-  })
-
-  return validateKeys(null, key, secret)
 }
 
 module.exports = async (ws, db) => {
   const { authPassword } = ws
 
   const [mainValid, paperValid] = await Promise.all([
-    validateCredentials(db, authPassword, 'main'),
-    validateCredentials(db, authPassword, 'paper')
+    getCredentialsState(db, 'main', authPassword),
+    getCredentialsState(db, 'paper', authPassword)
   ])
 
   send(ws, ['data.api_credentials.validation', {
